Introduce JSONMetadata, which can take a decoded JSON object directly
This is needed to support MD+BR deposit, where the metadata is embedded in
another JSON document (the MD+BR document).

JSON-based metadata formats are distinguished as we need to know that we can't
accept non-JSON-based formats in MD+BR deposits. See swordapp/swordv3#4 for the
upstream issue for this point to be folded into the spec.