Implement our own permissions, so we don't have to wait for ES to index
Previously, when a SWORDDeposit was created through the API, the client had to
wait a non-negliable amount of time for Invenio to index it into ElasticSearch
so that they could pass the permissions check when they subsequently attempt to
manipulate it. This overrides the invenio-deposit permissions check with a new
one that checks the database, so that there is no delay.